refactor: improved create user handler
Now when click on create user, it will clean all inputs in CreateUser component, it will return to the table and it will rerender the table. 🆒
Also added new button to create new dependencies (nothing functional yet). I'm thinking about making a whole new collection (let's hope it is small) where it stores the dependencies, so in the new button, it can add dependencies and it appears when creating a new user.