refactor: fixed updating an users info
Edit mode now works as intended, updated data is saved in a state in generalTable to then be filtered when the user clicks on save. The filter returns an array with the updated data, the row and columns affected, and sends a request to the backend to update the user. Finally. 😎