refactor: improved each users page
Added new service to search the token user in the backend.
The backend returns an object with the name of the user (if there are 2 names, it returns the first one). Then, it displays it in the 3 views.
Now, over the table, the 3 users can see their name displayed and their rol.