refactor: added automatically redirect to login
Now that the app config changed so the default route is simply '/', when starting the app nothing was rendered since the login component is in the route '/login'. To fix the eternal white screen, a new route was added in the Router, its path is '/' and automatically redirects to '/login'. Now it is possible to f5 the page without it breaking, and when starting the app, the user is automatically redirected to '/login'.